[squad] recvfuzz: Switch review fixes (cast + stale comment)
RecvFuzz Engineer (Switch) review of the frame-type generators found:
- IMMEDIATE_ACK frame-type write missing explicit (uint8_t) cast that
  every other frame-type byte write in the file uses (consistency).
- Stale multi-line comment in WriteShortHeaderPacket describing a
  now-superseded frame-selection approach.

Verified: /WX-clean Debug x64 schannel build; recvfuzz smoke run
(-timeout:8000 -seed:12345) exits 0.

Copilot-Session: d1480271-874a-495f-a1d6-9baa5795b616

+++ src/tools/recvfuzz/recvfuzz.cpp
@@ -971,7 +971,7 @@ void WriteImmediateAckFrame(
     if (*Offset + 1 > BufferLength) {
         return;
     }
-    Buffer[(*Offset)++] = QUIC_FRAME_IMMEDIATE_ACK;
+    Buffer[(*Offset)++] = (uint8_t)QUIC_FRAME_IMMEDIATE_ACK;
 }
 
 void WriteTimestampFrame(
@@ -1373,8 +1373,7 @@ void WriteShortHeaderPacket(
     uint16_t BufferSize = BufferLength - *HeaderLength;
     uint16_t PayloadLength = 0;
     
-    // Build frames (1-RTT packets typically contain STREAM frames, PING, etc.
-    // For fuzzing purposes, we'll use STREAM frames, ACK frames and padding)
+    // Build the requested frames.
     WriteFrames(&PayloadLength, BufferSize, Buffer + *HeaderLength, PacketParams);
     PayloadLength += GetRandom<uint8_t>(64); // More random padding
